* 'master' of https://bitbucket.org/gigapixxel/site_bookbox:
  Mudanças a pedido do cliente!
  mudanças de layout a pedido do cliente!
  Banner Blog Atualizado!
  Mudanças
  validação nos  populateSelectBox do /admin/prospection/course/from
  Banners Finalizados como pedido pelo cliente! Faixa Frete Gratis finalizada!
  Mudanças a pedido do cliente finalizadas!
  Cupom de desconto
  Trabalhando em Mudanças a pedido do cliente!
  Layout carrinho de compras finalizado 100%!
  Layout carrinho de compras em andamento!
  ok
  Carrinho de Compras

+++ database/migrations/2021_01_06_105456_create_city_table.php
@@ -7,30 +7,32 @@ use Illuminate\Support\Facades\Schema;
 class CreateCityTable extends Migration
 {
     /**
-     * Run the migrations.
-     *
-     * @return void
-     */
+    * Run the migrations.
+    *
+    * @return void
+    */
     public function up()
     {
-        Schema::create('city', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('country_id')->references('id')->on('country')->onDelete('cascade')->nullable();
-            $table->foreignId('state_id')->references('id')->on('state')->onDelete('cascade')->nullable();
-            $table->string('name', 150)->nullable();
-            $table->integer('created_by')->nullable();
-            $table->integer('updated_by')->nullable();
-            $table->integer('deleted_by')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('city')) {
+            Schema::create('city', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('country_id')->references('id')->on('country')->onDelete('cascade')->nullable();
+                $table->foreignId('state_id')->references('id')->on('state')->onDelete('cascade')->nullable();
+                $table->string('name', 150)->nullable();
+                $table->integer('created_by')->nullable();
+                $table->integer('updated_by')->nullable();
+                $table->integer('deleted_by')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
+    * Reverse the migrations.
+    *
+    * @return void
+    */
     public function down()
     {
         Schema::dropIfExists('city');
